packages/api/src/api/settings.py
import os

from dotenv import load_dotenv

load_dotenv()

REDIS_URL = os.getenv("REDIS_URL", "redis://localhost:6379")

if __name__ == "__main__":
    print(f"Loaded settings: REDIS_URL={REDIS_URL}")
